refactor: lowered include folders name
