+++ examples/tcpEcho/echoClient.cpp
@@ -1,5 +1,5 @@
 #include <iostream>
-#include "Polymorph/Network/tcp/Client.hpp"
+#include "polymorph/network/tcp/Client.hpp"
 #include "MessageDto.hpp"
