+++ include/polymorph/network/Packet.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "Polymorph/Network/PacketHeader.hpp"
+#include "polymorph/network/PacketHeader.hpp"
 
 namespace polymorph::network
 {
