+++ include/polymorph/network/dto/SessionTransferRequestDto.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "Polymorph/Network/types.hpp"
+#include "polymorph/network/types.hpp"
 
 namespace polymorph::network
 {
@@ -15,10 +15,9 @@ namespace polymorph::network
 #ifdef _WIN32
 #pragma pack(push, 1)
 #endif
-    struct SessionTransferResponseDto
+    struct SessionTransferRequestDto
     {
         static constexpr OpId opId = 2;
-        AuthorizationKey authKey;
     }
 #ifdef _WIN32
         ;
