+++ include/polymorph/network/exceptions/UnauthorizedException.hpp
