+++ include/polymorph/network/exceptions/UnknownAuthorizationKeyException.hpp
