+++ include/polymorph/network/tcp/Client.hpp
@@ -11,7 +11,7 @@
 #include <queue>
 #include <iostream>
 #include "APacketHandler.hpp"
-#include "Polymorph/Network/dto/ACKDto.hpp"
+#include "polymorph/network/dto/ACKDto.hpp"
 
 namespace polymorph::network::tcp
 {
