+++ include/polymorph/network/udp/Connector.hpp
@@ -11,7 +11,7 @@
 #include <asio/io_context.hpp>
 #include <queue>
 #include <vector>
-#include "Polymorph/Network/udp/IPacketHandler.hpp"
+#include "polymorph/network/udp/IPacketHandler.hpp"
 
 namespace polymorph::network::udp {
