+++ include/polymorph/network/udp/IPacketHandler.hpp
@@ -8,8 +8,8 @@
 #pragma once
 
 #include <asio/ip/udp.hpp>
-#include "Polymorph/Network/types.hpp"
-#include "Polymorph/Network/PacketHeader.hpp"
+#include "polymorph/network/types.hpp"
+#include "polymorph/network/PacketHeader.hpp"
 
 namespace polymorph::network::udp
 {
