+++ include/polymorph/network/udp/ServerPacketManager.hpp
@@ -11,7 +11,7 @@
 #include <map>
 #include <asio/steady_timer.hpp>
 #include <utility>
-#include "Polymorph/Network/types.hpp"
+#include "polymorph/network/types.hpp"
 
 namespace polymorph::network::udp
 {
