+++ src/src/SessionStore.cpp
@@ -6,12 +6,12 @@
 */
 
 
-#include "Polymorph/Network/SessionStore.hpp"
+#include "polymorph/network/SessionStore.hpp"
 #include "authorizationKey.hpp"
-#include "Polymorph/Network/exceptions/UnknownAuthorizationKeyException.hpp"
-#include "Polymorph/Network/exceptions/UnauthorizedException.hpp"
-#include "Polymorph/Network/exceptions/AlreadyRegisteredException.hpp"
-#include "Polymorph/Network/exceptions/UnknownSessionException.hpp"
+#include "polymorph/network/exceptions/UnknownAuthorizationKeyException.hpp"
+#include "polymorph/network/exceptions/UnauthorizedException.hpp"
+#include "polymorph/network/exceptions/AlreadyRegisteredException.hpp"
+#include "polymorph/network/exceptions/UnknownSessionException.hpp"
 
 void polymorph::network::SessionStore::registerAuthoredClient(asio::ip::udp::endpoint endpoint,
         polymorph::network::SessionId sessionId, polymorph::network::AuthorizationKey key)
